ChrstphrR 2004/04/21 (EDT) Common MapLoad changes
Bug Fix I introduced :P --  misspelt an index variable, and it would core the server when loading a shop.

AND ... inserted a reset of Lines in the script load process, so that
the scripts will show the current linenumber on the current file, instead of Lines being the cumulative count. (With 206 script files and counting, it will help the scripters track their errors quicker)

Dev changes:
Core: Common.pas
Scripts: Errors will be MUCH easier to find.
User changes:
Server: Script errors will be MUCH easier to find.


git-svn-id: svn+ssh://svn.code.sf.net/p/fusion/code/core/trunk@625 a2e91628-aa1e-412e-9966-34abe3d8d708