- 05-09-04: Main unit - MapLoad changed - no more char freezes. [ChrstphrR] - 05-09-04: ScriptValidated does checking for MapLoad. [ChrstphrR] - 05-09-04: ScriptValidated parses comment lines in scripts proper. [ChrstphrR] - 05-09-04: ScriptValidated parses comment lines in scripts proper. [ChrstphrR] - 05-09-04: ScriptValidated returns false for nonexistant commands/typosr. [ChrstphrR] - 05-09-04: ScriptValidated - fixed 'monster' parsing problem Colus noticed. [ChrstphrR]
The first cures the Character Freeze on script errors during the MapLoad phase.

2nd-4th fixes are improvements in the new validation routine that does the dirty work of MapLoad.

The last fix is for the 'monster' definitions in scripts - when insufficient tokens are in the 1st and 4th parts of a monster defition, they are padded with zeros. (Fix for an old EWeiss error bug in the code.)


git-svn-id: svn+ssh://svn.code.sf.net/p/fusion/code/core/trunk@671 a2e91628-aa1e-412e-9966-34abe3d8d708